# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "UsersController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Current user profile, preferences, roles, interests",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Update profile / onboarding / interests",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteMe_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Delete (soft) the current account",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/preferences": {
      "get": {
        "operationId": "UsersController_preferences_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Get preferences",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_updatePreferences_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Update preferences (radius, notifications)",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dealy API",
    "description": "Dealy production API — swipe-first, location-aware deals.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token"
      }
    },
    "schemas": {
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 80
          },
          "graduationYear": {
            "type": "number",
            "minimum": 1950,
            "maximum": 2100
          },
          "campusId": {
            "type": "string",
            "format": "uuid"
          },
          "onboardingCompleted": {
            "type": "boolean"
          },
          "interests": {
            "description": "Category slugs",
            "maxItems": 30,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdatePreferencesDto": {
        "type": "object",
        "properties": {
          "searchRadiusMiles": {
            "type": "number",
            "minimum": 1,
            "maximum": 50
          },
          "notificationsEnabled": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}